requirements.txt
llama-parse
llama-index-core
langchain
langchain-text-splitters
python-dotenv
pandas
tiktoken
openai
tqdm
chromadb
rank_bm25
sentence-transformers
torch
langchain-openai

